Reorganize storage module
